+++ app/src/main/java/kz/burhancakmak/aysoftmobile/Models/Clients/SatilanMalListesi.java
@@ -4,8 +4,8 @@ public class SatilanMalListesi {
 
     private String StokKodu;
     private String StokAdi;
-    private Integer Miktar;
-    private Integer Tutar;
+    private Float Miktar;
+    private Float Tutar;
 
     public String getStokKodu() {
         return StokKodu;
@@ -23,19 +23,19 @@ public class SatilanMalListesi {
         StokAdi = stokAdi;
     }
 
-    public Integer getMiktar() {
+    public Float getMiktar() {
         return Miktar;
     }
 
-    public void setMiktar(Integer miktar) {
+    public void setMiktar(Float miktar) {
         Miktar = miktar;
     }
 
-    public Integer getTutar() {
+    public Float getTutar() {
         return Tutar;
     }
 
-    public void setTutar(Integer tutar) {
+    public void setTutar(Float tutar) {
         Tutar = tutar;
     }
 }
